Refactor: Raspberry to ARM
