+++ tests/test_arm/conftest.py
@@ -6,9 +6,6 @@ import RPi.GPIO as gpio
 
 from unit.gpio import OperateGPIO as og
 
-# logging.getLogger('raspberry.pi3_gpio').setLevel(logging.CRITICAL)
-# logging.getLogger('unit.ping').setLevel(logging.CRITICAL)
-
 
 @pytest.fixture(scope="module")
 def rpi_gpio(my_pins):
